marked the test as Ignored
- because it requires a specific RabbitMQ configuration which
may not be available to all those who want to run the build locally.

It should not be deleted because it can still be useful to be run manually every once in a while -
say, with a newer RabbitMQ Client library version just to check that it still works for us.